+++ 02.wo_Defects/unlock_without_lock.c
@@ -116,8 +116,8 @@ void unlock_without_lock_002 ()
 {
 #if !defined(CHECKER_POLYSPACE)
 	pthread_t tid1,tid2;
-	int const unlock_without_lock_002_t1 = 10;
-	int const unlock_without_lock_002_t2 = 20;
+	intptr_t const unlock_without_lock_002_t1 = 10;
+	intptr_t const unlock_without_lock_002_t2 = 20;
 	pthread_mutex_init(&unlock_without_lock_002_glb_mutex, NULL);
 	pthread_create(&tid1, NULL, unlock_without_lock_002_tsk_001, (void *)unlock_without_lock_002_t1);
 	pthread_create(&tid2, NULL, unlock_without_lock_002_tsk_001, (void *)unlock_without_lock_002_t2);
@@ -209,7 +209,7 @@ void unlock_without_lock_003 ()
 {
 #if ! defined(CHECKER_POLYSPACE)
 	pthread_t tid1,tid2;
-	int t1 = 10, t2 = 20;
+	intptr_t t1 = 10, t2 = 20;
 	pthread_mutex_init(&unlock_without_lock_003_glb_mutex, NULL);
 	pthread_create(&tid1, NULL, unlock_without_lock_003_tsk_001, (void *)t1);
 	pthread_create(&tid2, NULL, unlock_without_lock_003_tsk_001, (void *)t2);
@@ -292,8 +292,8 @@ void unlock_without_lock_004 ()
 {
 #if ! defined(CHECKER_POLYSPACE)
 	pthread_t th1,th2;
-	long int t1 = 10;
-	long int t2 = 20;
+	intptr_t t1 = 10;
+	intptr_t t2 = 20;
     pthread_create(&th1, NULL, unlock_without_lock_004_tsk_001, (void *)t1);
     pthread_create(&th2, NULL, unlock_without_lock_004_tsk_002, (void *)t2);
 	sleep(1);
@@ -404,8 +404,8 @@ void unlock_without_lock_006 ()
 	int thread_set = 0;
 #if ! defined(CHECKER_POLYSPACE)
 	pthread_t th1,th2;
-    long int t1 = 10;
-	long int t2 = 20;
+        intptr_t t1 = 10;
+	intptr_t t2 = 20;
 	if ( thread_set == CREATE_THREAD)
 	{
 	   ;
